fix(mobile): gate Join on live close state, not fetch-time isJoinable
CodeRabbit, valid — a real correctness gap. The Join affordances (confirm
sheet, invite screen) gated on isJoinable as captured at fetch time, but
closesAtUtc can elapse WHILE the sheet or screen is open. JoinClosesLabel
already recomputes live with a timer; the buttons did not, so Join could stay
enabled past the deadline and submit a join the BE would reject.

Extracted useLiveJoinState — the ticking join-status logic that was inline in
JoinClosesLabel — as the single source of truth. JoinClosesLabel now consumes
it (its inline timer removed); the confirm sheet disables Join (title -> Closed)
when the live state is closed; the invite screen branches its closed-state on
the live state instead of the fetch-time isJoinable flag. A league that closes
mid-view now blocks the join on both surfaces.

Docs: revised the intro + gap narrative to past tense — Phase 1 (find & join)
has shipped; the contracts remain the Phase 2 reference.

Validation: tsc --noEmit clean; 93 jest tests pass (the joinDisplay contract
tests already cover the closed/countdown/date states this hook returns).
